Fixed some tests.| #1907
